move events attaching to use-swr
